+++ doc/conf.py
@@ -86,7 +86,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = 'Wpull'
-copyright = '2013-2014, Christopher Foo'
+copyright = '2013-2015, Christopher Foo and others'
 
 
 # The version info for the project you're documenting, acts as replacement for