+++ wpull/__init__.py
@@ -1,1 +1,2 @@
 '''Wget-compatible web downloader and crawler.'''
+# Wpull. Copyright 2013-2015: Christopher Foo and others. License: GPL v3.