+++ wpull/urlfilter.py
@@ -162,15 +162,22 @@ class RecursiveFilter(BaseURLFilter):
 
 class LevelFilter(BaseURLFilter):
     '''Allow URLs up to a level of recursion.'''
-    def __init__(self, max_depth):
+    def __init__(self, max_depth, inline_max_depth=5):
         self._depth = max_depth
+        self._inline_max_depth = inline_max_depth
 
     def test(self, url_info, url_table_record):
-        if url_table_record.inline:
-            return True
+        if self._inline_max_depth and url_table_record.inline and \
+                url_table_record.inline > self._inline_max_depth:
+            return False
 
         if self._depth:
-            return url_table_record.level <= self._depth
+            if url_table_record.inline:
+                # Allow exceeding level to allow fetching html pages with
+                # frames, for example, but no more than that
+                return url_table_record.level <= self._depth + 2
+            else:
+                return url_table_record.level <= self._depth
         else:
             return True